Port SimpBMS bug fixes from BentlyjagXJS-fellten-battery
Both branches independently implemented the SimpBMS current limits and the
calculated road speed. The fellten branch subsequently debugged its copies;
this branch did not. Ports those fixes across.

Fix the current limits being 10x too small. The 0.1A CCL/DCL digits were
converted to amps at decode and then scaled by 0.1 again in Task100Ms, so a
300A DCL became a 30A limit. Store raw digits as int, matching fellten, and
scale exactly once. ChargeAllowed's zero test becomes < 5 digits to preserve
its original 0.5A intent.

Move the BMS timeout reset from 0x373 to 0x351. The Orion in this setup does
not send 0x373, so timeoutCounter never reset, BMSDataValid() was permanently
false and the limit application above was silently skipped entirely.

Zero BMS_Vmin/Vmax/Tmin/Tmax on timeout instead of leaving the last known
cell voltages and temperatures on display after a CAN dropout. idcmax/idcmin
stay inside the valid branch so a dropout leaves the configured limits in
force and cannot cut drive.

Decode 0x356 pack voltage as 0.1V per digit, not 0.01V, matching the 0.1A
current scaling in the same frame.

Bump VER to 2.41.A. The web UI caches the param JSON keyed on the version
string, and the CAT_SPEEDO category was added without bumping it, so the
Speedometer params never appeared.

Narrow HeatPotDir max from 4 to 3; ABOVEBELOW only defines values 0-3.

Also take fellten's fixed GS450H gear ratio. This car runs a Leaf motor so
GetGearboxRatio returns at the inverter check and the GS450H path is
unreachable here, but it keeps the two branches aligned.

+++ include/simpbms.h
@@ -35,10 +35,9 @@ public:
 private:
   bool BMSDataValid();
   bool ChargeAllowed();
-  // CCL and DCL are sent by Simp as 0.1A per digit and scaled to amps on
-  // decode, so both of these are stored in amps.
-  float chargeCurrentLimit = 0;
-  float dischargeCurrentLimit = 0;
+  // CCL and DCL are sent by Simp as 0.1A per digit and stored the same way.
+  int chargeCurrentLimit = 0;
+  int dischargeCurrentLimit = 0;
   int timeoutCounter = 0;
   float minCellV = 0;
   float maxCellV = 0;
